rename: Core 폴더 이동
