+++ src/main/java/prography/cakeke/server/store/domain/StoreAndTag.java
@@ -7,7 +7,7 @@ import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import prography.cakeke.server.common.Core;
+import prography.cakeke.server.common.domain.Core;
 
 @Entity
 @Getter
