+++ src/main/java/prography/cakeke/server/store/domain/StoreTag.java
@@ -9,7 +9,7 @@ import jakarta.persistence.Enumerated;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import prography.cakeke.server.common.Core;
+import prography.cakeke.server.common.domain.Core;
 
 @Entity
 @Getter
